flask-sql-example-main/requirements.txt
flask
flask_cors
flask-limiter
flask_sqlalchemy
pyotp
